refactor: change input process to be easy to test
