+++ src/main/java/lotto/parser/InputParser.java
@@ -1,0 +1,5 @@
+package lotto.parser;
+
+public interface InputParser<T> {
+    T parse(String input);
+}
