--- src/main/java/lotto/view/converter/InputConverter.java
@@ -1,5 +1,0 @@
-package lotto.view.converter;
-
-public interface InputConverter<T> {
-    T convert();
-}
